refact(calculator): Calculators in separate packages
The former `calculator.py` was getting a bit crowded, so it got separated into
`numpy_`, `tensorflow_` and `jax_` packages.
