+++ piquasso/_backends/sampling/simulator.py
@@ -16,7 +16,7 @@
 from ..simulator import BuiltinSimulator
 from piquasso.instructions import preparations, gates, measurements, channels
 
-from piquasso._backends.calculator import NumpyCalculator, JaxCalculator
+from piquasso._backends.calculators import NumpyCalculator, JaxCalculator
 
 from .state import SamplingState
 from .calculations import (
