+++ piquasso/decompositions/clements.py
@@ -31,7 +31,7 @@ from piquasso.api.calculator import BaseCalculator
 
 from piquasso._math.indices import get_operator_index
 
-from piquasso._backends.calculator import NumpyCalculator
+from piquasso._backends.calculators import NumpyCalculator
 
 if TYPE_CHECKING:
     import numpy as np
