+++ tests/_math/test_decompositions.py
@@ -26,7 +26,7 @@ from piquasso._math.decompositions import (
     decompose_to_pure_and_mixed,
 )
 
-from piquasso._backends.calculator import NumpyCalculator, TensorflowCalculator
+from piquasso._backends.calculators import NumpyCalculator, TensorflowCalculator
 
 
 @pytest.mark.parametrize("calculator", [NumpyCalculator(), TensorflowCalculator()])
